refactor: Move loading and releasing sound files into a new hook
