+++ utils/sound/getSoundsKey.ts
@@ -1,0 +1,1 @@
+export const getSoundsKey = (hz: number) => `sound${hz}hz`;
